Add index to ar_power_of_attorney_requests.redacted_at (#22095)
This commit adds an index to the `redacted_at` column of the `ar_power_of_attorney_requests` table.

Adding this index will improve the performance of queries that filter or order by the `redacted_at` column, particularly when retrieving redacted power of attorney requests.

The index is created concurrently to avoid locking the table during the operation. It also includes a check to prevent attempting to add the index if it already exists.

Co-authored-by: OJ Bucao <9256675+ojbucao@users.noreply.github.com>