feat(AccreditedRepresentativePortal): Expose user LOA in API response (#22221)
This commit updates the `RepresentativeUsersController` within the
Accredited Representative Portal to include the current user's
Level of Assurance (LOA) in the API response.

The `loa` attribute is now part of the user object returned by
the `/accredited_representative_portal/v0/user` endpoint.

The corresponding request specs in `user_spec.rb` have been
updated to reflect this change and ensure the `loa` attribute
is correctly serialized.

Co-authored-by: OJ Bucao <9256675+ojbucao@users.noreply.github.com>