[ARP] Surface Benefits Intake UUID (#22963)
* [ARP] Surface Benefits Intake UUID
Rather than our own internal ID, for claim submissions. We learned that representative do have help resources where this ID is actually useful. Apparently Central Mail Portal operators are able to make use of Benefits Intake UUID as well as Central Mail's Packet ID. And the help resources available to representatives can make contact with those operators directly or indirectly.

This ties into a UX concern. It would potentially be better UX for a length claim submission to happen asynchronously such that this ID would not be available immediately to display on the form submission confirmation.

For now, we're going to do synchronous submission.

* lint fix