API-48176-pdf-constructor-claimant-name-v2-fix (#23185)
At the moment POA v2 does not support dependent claimants.  One issue this caused is that v2 poa work was built to match the v1 POA schema.  However the claimant's names (first and last) are not present in the form data like they are in v1 so the areas in v2 where we were grabbing those values similar to v1 will not work (are not working).

This PR fixes handling of claimant names in the org PDF constructor and the v2 poa_form_builder_job so that the correct values can be found.

- This does not fix all areas just what is required for the current scope of the work.  There is a follow up ticket for the remainder of the refactoring/fix work here: API-48176