Implement Disability Compensation monitor for SavedClaim errors
Creates a method in the DisabilityCompensation::Loggers::Monitor class we can use to log the details of ActiveRecord save errors when a SavedClaim::DisabilityCompensation::Form526AllClaim fails to save. This method leverages the Logging::BaseMonitor framework implemented by the Pensions and Burials team which provides a standardized way to log useful information and increment an associated StatsD metric key.

Note we are including the InProgressForm ID for the associated SavedClaim in this log; this is useful because we can inspect the in progress form data in the production console to understand the selections a Veteran made, even if we didn't successfully create a SavedClaim object. But ideally we should get all the information we need from the logged Active Record errors