Api 48861 decide response refactor (#23590)
* Move validate_decide_representative_params! into decide service
* Moves the method into the new decide service file
* Adjusts related tests, adds tests into decide_spec
* Moves get_poa_request into Decide service file
* Moves the get_poa_request into the decide service file, no longer borrowing from the Show service
* Adds a test and adjusts related tests
* Moves build_target_veteran into the decide service
Just another attempt to clean up the controller.  Having these built veteran values will allow us to add
to the response in a much cleaner fashion using the Decide service file.  This is a pattern I wish I had
noticed earlier in the accepted work, it would have helped make that even cleaner but stopping here for now
to keep the scope related to the response update needed
* Moves build target veteran into decide service
* small clean up on variable name which was already used in the controller
* Adds tests and adjusts any related tests