[Form 526] Improve logging detail on Veteran claim submission (#23557)
* Implement Disability Compensation monitor for SavedClaim errors

Creates a method in the DisabilityCompensation::Loggers::Monitor class we can use to log the details of ActiveRecord save errors when a SavedClaim::DisabilityCompensation::Form526AllClaim fails to save. This method leverages the Logging::BaseMonitor framework implemented by the Pensions and Burials team which provides a standardized way to log useful information and increment an associated StatsD metric key.

Note we are including the InProgressForm ID for the associated SavedClaim in this log; this is useful because we can inspect the in progress form data in the production console to understand the selections a Veteran made, even if we didn't successfully create a SavedClaim object. But ideally we should get all the information we need from the logged Active Record errors

* Implement new SavedClaim error logging on submit_all_claim endpoint

Adds the new track_saved_claim_save_error method call to the point where a SavedClaim fails to save, with simple tests that the method is called.

* Add more specific conditions to logging test and fix bug

Fleshed out the argument tests for the new logging and discovered a bug where the default form id for our SavedClaims is mismatched with what we have for InProgressForms

* Refactor logging payload to use whitelisted keys

As originally implemented, this new logging behavior would have filtered out useful error information via the parameter sanitization in config/initializers/filter_parameter_logging.rb; this behavior is intended to prevent PII from ending up in the logs, but I'm logging model validation errors which _shouldn't_ contain PII as far as I can tell.

Refactors the logging to use the correct whitelisted keys so the data gets logged, and had to make some changes to ensure the serialization behavior works correctly with the ActiveModel::Error objects I anticipate will be present here.

* Small cleanup

* Fix linting errors

* Add begin ensure block to make sure we log something if the IPF lookup fails

* Fix another linting error

* PR Feedback: use different syntax for Flipper clauses

* Correct missed flipper line

---------

Co-authored-by: Rachal Cassity <rachal.cassity@oddball.io>