Disable Lighthouse banking info lookup on Form 526 submit
Adds a Flipper to block retrieving banking info from Lighthouse if the user did not provide it in their form. This call was failing during Lighthouse outages and it was blocking submission. The decision was made to remove it entirely, since the Veteran left that part of the form blank and we shouldn't fill it out for them without their knowledge.

All code making this call will be removed when the Flipper is removed after a successful launch in production