Bypassed default values for checkboxes when blank on several forms (#24388)
* Removed default values for checkboxes on several forms

These checkboxes are now left blank when there is no data to populate them.  The forms modified are 10-10d (both regular and merged) and the 10-7959c revision used in the merged form flow.

* Cleaned up logic

Field applicant_secondary_eob is not present on this form, keying off of applicant_secondary_provider instead. The has_other_health_insurance check is redundant now that we're looking for the presence of a blank section.

* Revert "Cleaned up logic"

This reverts commit f1ce64f219c3e03a3ed162a89593965e321dde89.