[Unblock Form 526 Submission] Remove call to retrieve Veteran banking data if it is not supplied (#24481)
* Disable Lighthouse banking info lookup on Form 526 submit

Adds a Flipper to block retrieving banking info from Lighthouse if the user did not provide it in their form. This call was failing during Lighthouse outages and it was blocking submission. The decision was made to remove it entirely, since the Veteran left that part of the form blank and we shouldn't fill it out for them without their knowledge.

All code making this call will be removed when the Flipper is removed after a successful launch in production

* Add banking info presence loggers

Creates monitor log methods for Form 526 submissions with or without banking info. We can use these to monitor how often Veterans are submitting without providing this information, in case we notice an abnormally large number which may indicate a problem

* Add logging for submissions with and without banking info

Adds new monitor log method calls for Form 526 submissions made with and without banking information. We can use these to monitor how often Veterans are submitting without providing this information (which is allowed), in case we notice an abnormally large number which may indicate a problem with prefilling that information or an issue with the form itself

* Pre-PR cleanup

* CI-caught fixes

* PR Feedback: hide new logging behind flipper as well

* PR Feedback: add logging tests for StatsD metric increment in logger

* Remove unnecessary VCR block from test with no external call made

* Fix test that broke due to logging changes on master

---------

Co-authored-by: Rebecca Tolmach <10993987+rmtolmach@users.noreply.github.com>