Letter ready email job retry environment variable and config update.
What is this?
We are updating the email job retry logic to use a environment variable.

Why do we need this?
Staging is getting clogged up with retry jobs because of test cases. We want to provide a way to turn down the number of retries on staging.

How did we solve this?
Added to the config settings. Used the settings in a already existing constant. Added a default value to fall back if the value does not exist in an environment.