Background: - MAX_EMAIL_ATTEMPTS sets the maximum number of email tries across jobs. - SIDEKIQ_RETRY_OPTIONS sets the retries for each job. - The jobs being run:   - LetterReadyEmailJob - Triggered when letter is ready   - LetterReadyRetryEmailJob - Triggered when temporary-failure is received from VA notify - In both jobs we check max email attempts, so if 5 is set in the environment variable email retry will only occur five times across both jobs. - Case 1:   - If failures occur in LetterReadyEmailJob it willl retry there five times - Case 2:   - If VA notify sends temporary-failure after two LetterReadyEmailJobs retries LetterReadyRetryEmailJob will be triggered.   - If LetterReadyRetryEmailJob continues to fail it will use the remaining tries. In this case 3.
Why do we need this?
Sets the max retry count across jobs and also for each job. Allows us to lower the amount of retries per environment. On prod, and lower environments the 16 retry count value is too high. Was flagged as a risk to the prod db.

What did we do?
Created a shared variable between sidekiq retries and max email attempts. Allows us to configure per environment, and adjust this value easily in the future if we need to.