Update the retry count on letter ready email job. (#24831)
* Letter ready email job retry environment variable and config update.

What is this?
We are updating the email job retry logic to use a environment variable.

Why do we need this?
Staging is getting clogged up with retry jobs because of test cases. We want to provide a way to turn down the number of retries on staging.

How did we solve this?
Added to the config settings. Used the settings in a already existing constant. Added a default value to fall back if the value does not exist in an environment.

* No need to have a default.

* Safe navigate and default.

* Added two new tests that will be changed to add the new environment variable tests.

* Background:
- MAX_EMAIL_ATTEMPTS sets the maximum number of email tries across jobs.
- SIDEKIQ_RETRY_OPTIONS sets the retries for each job.
- The jobs being run:
  - LetterReadyEmailJob - Triggered when letter is ready
  - LetterReadyRetryEmailJob - Triggered when temporary-failure is received from VA notify
- In both jobs we check max email attempts, so if 5 is set in the environment variable email retry will only occur five times across both jobs.
- Case 1:
  - If failures occur in LetterReadyEmailJob it willl retry there five times
- Case 2:
  - If VA notify sends temporary-failure after two LetterReadyEmailJobs retries LetterReadyRetryEmailJob will be triggered.
  - If LetterReadyRetryEmailJob continues to fail it will use the remaining tries. In this case 3.

Why do we need this?
Sets the max retry count across jobs and also for each job. Allows us to lower the amount of retries per environment. On prod, and lower environments the 16 retry count value is too high. Was flagged as a risk to the prod db.

What did we do?
Created a shared variable between sidekiq retries and max email attempts. Allows us to configure per environment, and adjust this value easily in the future if we need to.

* More specific variable name for the module.

* Why are we making this change?
In collaboration with the event bus team and after into digging into how the system works we decided this was better handled by a constant hard coded value instead of a environment variable. We will lower the retries in all enviornments and hard code the values. The retry job should be lower than the first email job because there are going to be less tries available for that job.

* Constant updates.

* Remove environment variables tests now that we are using constants.

* Retry count test for letter ready email job.

* Revert gemfile.lock

* Remove dated comment. We adjusted the business rules.

* Test around the new letter ready email retry job business rule for retries.

* End to end test around max retry variable.

* Add a test for when there is sidekiq jobs left but we reached max retry.

* Out of attempts test.

* Tests for max emails setting.

* Working notify email status callback tests.

* Working limit end to end tests.

* Revert gemfile.lock

* Clean up tests for PR.

* Add comments for the environment variables.

* Update outdated comment.

* Remove business rule heading on unit test.

* Remove business rule heading to tests.

* Check for sheduled retry job call on 4 and 5 attempts for vanotify email status callback.

* Space between retries.

* Remove the already tested code.

* Keep test that explain context around current max email setting.

* Update tests to use constant, keep test context around why the constant is expected to be this value.

* Update test to use a constant, but keep context on why test exists.

* Update descriptions to use constant.

* Comment updates and use of constant.

* Change to use constant value, not a static value.